+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@
 - add support for automatically computed parameter values
 - add live command display during execution
 - add support for Gradle test suites
+- add support for read-only ingredient files (`ingredient_files`)
 
 **Changed**